# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Freshify API Documentation
  description: |-
    An API for creating and managing a household inventory and shopping list.
    The API is written in Java using the Spring Boot framework. The API uses a
    MySQL database. The API is documented using Swagger. The API is hosted using
    Apache.
    Some useful links:
    - [The GitLab repository](https://gitlab.stud.idi.ntnu.no/idatt2106-v23-10/smartmat-backend) 
  contact:
    email: vkbugge@hotmail.com
  license:
    name: MIT License
    url: 'https://opensource.org/licenses/MIT'
  version: 1.0.0
servers:
  - url: http://localhost:8080
...
paths:
  /user:
    post:
      tags:
        - account
      summary: Create a user
      description: Creates a user
      operationId: createUser
      requestBody:
        description: The user to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFull'
        '409':
          description: User already exists
    put:
      tags:
        - account
      summary: Update user
      description: Updates a user
      operationId: updateUser
      requestBody:
        description: The user to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFull'
        '404':
          description: User not found
  /user/{id}:
    get:
      tags:
        - account
      summary: Get user by id
      description: Returns a single user
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID of user to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFull'
        '404':
          description: User not found
    delete:
      tags:
        - account
      summary: Delete user
      description: Deletes a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID of user to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: successful operation
        '401':
          description: Access denied
        '404':
          description: User not found
  /user/{id}/households:
    get:
      tags:
        - household
        - account
      summary: Get a user's households
      description: Gets a list of households of a given user
      operationId: getHouseholds
      parameters:
        - name: id
          in: path
          description: ID of user to get households from
          required: true
          schema:
            type: integer
            format: int64
      responses:
          '200':
            description: successful operation
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Household'
          '401':
            description: Access denied
          '404':
            description: User not found
  /user/login:
    post:
      tags:
        - account
      summary: Login user
      description: Logs in a user
      operationId: loginUser
      requestBody:
        description: The user to be logged in
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserFull'
        '401':
          description: User could not be logged in
...
  /household/{id}/users/{userId}:
    delete:
      tags:
        - household
      summary: Remove a user from a household
      description: Removes a user from a household. Can only be done by a superuser or the user themselves
      operationId: removeUserFromHousehold
      parameters:
        - name: id
          in: path
          description: ID of household to delete a user from
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          description: ID of user to delete from the household
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: successful operation
        '404':
          description: User not found/Household not found/User does not belong to household
        '401':
          description: Access denied
...
components:
  schemas:
    CreateUser:
      type: object
      properties:
        email:
          type: string
          format: email
          example: jens@gmail.com
        password:
          type: string
          format: password
          example: password123
        firstName:
          type: string
          example: John
    LoginUser:
      type: object
      properties:
        email:
          type: string
          format: email
          example: lennon@gmail.com
        password:
          type: string
          format: password
          example: hunter2
    UpdateUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          format: email
          example: jens@gmail.com
        password:
          type: string
          format: password
          example: password123
        firstName:
          type: string
          example: John
    UserFull:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        email:
          type: string
          format: email
          example: example@example.org
        firstName:
          type: string
          example: John
...
    Household:
      type: object
      properties:
        id:
          type: integer
          example: 123
        name:
          type: string
          example: "My family"